Fixed proficiencies being floats when displayed in the lore.

+++ src/main/java/com/stelios/cakenaysh/Util/BattleItemBuilder.java
@@ -683,21 +683,21 @@ public class BattleItemBuilder extends ItemBuilder{
         if (this.getStat("armorProficiency") != 0){
             loreList.add(0, Component.text("Armor Proficiency: ", TextColor.color(200, 200, 200))
                     .decoration(TextDecoration.ITALIC, false)
-                    .append(Component.text(this.getStat("armorProficiency"), TextColor.color(240, 40, 50)))
+                    .append(Component.text((int) this.getStat("armorProficiency"), TextColor.color(240, 40, 50)))
                     .decoration(TextDecoration.ITALIC, false));
         }
 
         if (this.getStat("rangedProficiency") != 0){
             loreList.add(0, Component.text("Ranged Proficiency: ", TextColor.color(200, 200, 200))
                     .decoration(TextDecoration.ITALIC, false)
-                    .append(Component.text(this.getStat("rangedProficiency"), TextColor.color(240, 40, 50)))
+                    .append(Component.text((int) this.getStat("rangedProficiency"), TextColor.color(240, 40, 50)))
                     .decoration(TextDecoration.ITALIC, false));
         }
 
         if (this.getStat("meleeProficiency") != 0){
             loreList.add(0, Component.text("Melee Proficiency: ", TextColor.color(200, 200, 200))
                     .decoration(TextDecoration.ITALIC, false)
-                    .append(Component.text(this.getStat("meleeProficiency"), TextColor.color(240, 40, 50)))
+                    .append(Component.text((int) this.getStat("meleeProficiency"), TextColor.color(240, 40, 50)))
                     .decoration(TextDecoration.ITALIC, false));
         }
